pyproject.toml
[tool.poetry]
name = "legalops_4"
version = "0.1.0"
description = "legalsops for retriever and reader"
authors = ["Victor Venegas <dextex@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
flask = "^3.0.3"
flask-cors = "^5.0.0"
pymongo = "^4.8.0"
langchain-voyageai = "^0.1.1"
pytest = "7.1"
dspy-ai = "2.4.13"
python-dotenv = "^1.0.1"
anthropic = "^0.34.2"
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl"}
langchain = "^0.3.0"
pytest-asyncio = "0.21.0"
azure-storage-blob = "^12.23.1"
azure-identity = "^1.18.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
